+++ Core/Exporters/DelimiterSeparatedValueExporter.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Collections.Generic;
+using System.Text;
 
 namespace CKAN.Exporters
 {
@@ -20,7 +21,7 @@ namespace CKAN.Exporters
 
         public void Export(RegistryManager manager, IRegistryQuerier registry, Stream stream)
         {
-            using (var writer = new StreamWriter(stream))
+            using (var writer = new StreamWriter(stream, Encoding.UTF8, 4096, true))
             {
                 writer.WriteLine(string.Join(_delimiter,
                                              "identifier",