+++ Tests/Core/IO/CKANPathUtils.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using CKAN;
 using CKAN.IO;
 
-namespace Tests.Core
+namespace Tests.Core.IO
 {
     [TestFixture]
     public class CKANPathUtilsTests